it works. might cry.
Updated the <autologon> count to 5 - found out that it modifies the defaultpassword reg key and the autologon reg key based on the logoncount number. Set it to 5 so it will login 4 times, there's an extra one because 24H2 apparently somehow logs in once before it hits the main login screen and builds the user profile the first time. It subtracts 1 count each time it auto logs on after restart built into each script parts 1 through 4, and when it reaches part 4 on the 5th logon, it automatically DELETES the defaultpassword reg key and disabled autologon by changed the value from 1 to 0. Works beautifully. May cry as I spent 6 hours or more working on reg keys in the individual scripts and the XML file for this to work. UGH.